Merge commit 'cb0f5ed8111d4305cf6798aefdb4f9623570c83a'

* commit 'cb0f5ed8111d4305cf6798aefdb4f9623570c83a':
  Remove compiler warnings when building Bionic.

+++ libc/regex/regexec.c
@@ -153,7 +153,7 @@ regexec(const regex_t *preg, const char *string, size_t nmatch,
 		return(REG_BADPAT);
 	eflags = GOODFLAGS(eflags);
 
-	if (g->nstates <= CHAR_BIT*sizeof(states1) && !(eflags&REG_LARGE))
+	if (g->nstates <= (int)(CHAR_BIT*sizeof(states1)) && !(eflags&REG_LARGE))
 		return(smatcher(g, (char *)string, nmatch, pmatch, eflags));
 	else
 		return(lmatcher(g, (char *)string, nmatch, pmatch, eflags));
